+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProjectFileIT.java
@@ -36,6 +36,11 @@ public class ProjectFileIT extends AbstractServerIT {
     }
 
     @Test(timeout = 30000)
+    public void testSingleProfileUsingConfiguration() throws Exception {
+        simpleTest("projectfile/singleprofilecfg", ".*Hello, world.*", ".*54321.*");
+    }
+
+    @Test(timeout = 30000)
     public void testExternalProfile() throws Exception {
         simpleTest("projectfile/externalprofile", ".*Hello, world.*");
     }